+++ smartsheet/sheets.py
@@ -17,12 +17,12 @@
 
 from __future__ import absolute_import
 
-from .models.column import Column
-from .models.row import Row
-from .types import TypedList
 import logging
 import os.path
 from datetime import datetime
+from .models.column import Column
+from .models.row import Row
+from .types import TypedList
 from .util import deprecated
 from . import fresh_operation
 
@@ -42,11 +42,12 @@ class Sheets(object):
         Args:
             sheet_id (int): Sheet ID
             list_of_columns (list[Column]): One or more
-                Column objects.
+                Column objects
 
         Returns:
             Result
         """
+
         if isinstance(list_of_columns, (dict, Column)):
             arg_value = list_of_columns
             list_of_columns = TypedList(Column)
@@ -79,7 +80,7 @@ class Sheets(object):
 
         Args:
             sheet_id (int): Sheet ID
-            list_of_rows (list[Row]): An array of Row objects with the following attributes:
+            row_or_list_of_rows (list[Row]): An array of Row objects with the following attributes:
 
                One or more location-specifier attributes (optional)
 
@@ -184,7 +185,7 @@ class Sheets(object):
         return response
 
     def copy_rows(self, sheet_id, copy_or_move_row_directive_obj,
-                  include=None, ignore_rows_not_found=False):
+                  include=None, ignore_rows_not_found=None):
         """Copies Row(s) from the specified Sheet to the bottom of another
         Sheet.
 
@@ -724,7 +725,7 @@ class Sheets(object):
         return response
 
     def move_rows(self, sheet_id, copy_or_move_row_directive_obj,
-                  include=None, ignore_rows_not_found=False):
+                  include=None, ignore_rows_not_found=None):
         """Moves Row(s) to the bottom of another Sheet.
 
         Up to 5,000 row IDs can be specified in the request, but if
@@ -922,7 +923,7 @@ class Sheets(object):
 
         return response
 
-    def share_sheet(self, sheet_id, share_obj, send_email=False):
+    def share_sheet(self, sheet_id, share_obj, send_email=None):
         """Share the specified Sheet.
 
         Share the specified Sheet with the specified Users and
@@ -941,8 +942,8 @@ class Sheets(object):
         _op = fresh_operation('share_sheet')
         _op['method'] = 'POST'
         _op['path'] = '/sheets/' + str(sheet_id) + '/shares'
-        _op['query_params']['sendEmail'] = send_email
         _op['json'] = share_obj
+        _op['query_params']['sendEmail'] = 'true'
 
         expected = ['Result', 'Share']
 
@@ -1535,6 +1536,71 @@ class Sheets(object):
         _op['json'] = sort_specifier_obj
 
         expected = 'Sheet'
+
+        prepped_request = self._base.prepare_request(_op)
+        response = self._base.request(prepped_request, expected, _op)
+
+        return response
+
+    def import_csv_sheet(self, file, sheet_name=None, header_row_index=None, primary_column_index=None ):
+        """Imports a sheet.
+
+        Args:
+            file (string): path to CSV file.
+            sheet_name (string): destination sheet name
+            header_row_index (int): index (0 based) of row to be used for column names
+            primary_column_index (int): index (0 based) of primary column
+
+        Returns:
+            Result
+            """
+        if not all(val is not None for val in ['folder_id', 'file']):
+            raise ValueError(
+                ('One or more required values '
+                 'are missing from call to ' + __name__))
+
+        return self._import_sheet(file, "text/csv", sheet_name, header_row_index, primary_column_index)
+
+    def import_xlsx_sheet(self, file, sheet_name=None, header_row_index=None, primary_column_index=None ):
+        """Imports a sheet.
+
+        Args:
+            file (string): path to XLSX file.
+            sheet_name (string): destination sheet name
+            header_row_index (int): index (0 based) of row to be used for column names
+            primary_column_index (int): index (0 based) of primary column
+
+        Returns:
+            Result
+            """
+        if not all(val is not None for val in ['folder_id', 'file']):
+            raise ValueError(
+                ('One or more required values '
+                 'are missing from call to ' + __name__))
+
+        return self._import_sheet(file, "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+                                  sheet_name, header_row_index, primary_column_index)
+
+    def _import_sheet(self, file, file_type, sheet_name, header_row_index, primary_column_index):
+        """Internal function used to import sheet"""
+
+        if sheet_name is None:
+            head, tail = os.path.split(file)
+            sheet_name = tail or os.path.basename(head)
+
+        _data = open(file, 'rb').read()
+
+        _op = fresh_operation('import_sheet_into_folder')
+        _op['method'] = 'POST'
+        _op['path'] = '/sheets/import'
+        _op['headers'] = {'content-type': file_type,
+                          'content-disposition': 'attachment'}
+        _op['form_data'] = _data
+        _op['query_params']['sheetName'] = sheet_name
+        _op['query_params']['headerRowIndex'] = header_row_index
+        _op['query_params']['primaryColumnIndex'] = primary_column_index
+
+        expected = ['Result', 'Sheet']
 
         prepped_request = self._base.prepare_request(_op)
         response = self._base.request(prepped_request, expected, _op)